Move Telemetry to helpers for cleanliness
