+++ simgui-window.json
@@ -41,7 +41,7 @@
     "###Joysticks": {
       "Collapsed": "0",
       "Pos": "349,627",
-      "Size": "976,278"
+      "Size": "976,272"
     },
     "###NetworkTables": {
       "Collapsed": "0",
