+++ src/main/java/frc/robot/helpers/Telemetry.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.helpers;
 
 import com.ctre.phoenix6.SignalLogger;
 import com.ctre.phoenix6.Utils;
